Superstore Sales Analysis
The sales analysis is done in Power BI it contains the pie charts which depict the profit by region, Category, and Segment. It also holds the matrix which says the sales by region the line chart to identify the trend of profit in the year. The cue cards show the sales, Profit, and Quantity in a year. the report is filtered by using the States.